// src/domain/contracts/repos/schema.ts
export type TCodUfIbge =
  | '11'
  | '12'
  | '13'
  | '14'
  | '15'
  | '16'
  | '17'
  | '21'
  | '22'
  | '23'
  | '24'
  | '25'
  | '26'
  | '27'
  | '28'
  | '29'
  | '31'
  | '32'
  | '33'
  | '35'
  | '41'
  | '42'
  | '43'
  | '50'
  | '51'
  | '52'
  | '53';
export type TUfEmi =
  | 'AC'
  | 'AL'
  | 'AM'
  | 'AP'
  | 'BA'
  | 'CE'
  | 'DF'
  | 'ES'
  | 'GO'
  | 'MA'
  | 'MG'
  | 'MS'
  | 'MT'
  | 'PA'
  | 'PB'
  | 'PE'
  | 'PI'
  | 'PR'
  | 'RJ'
  | 'RN'
  | 'RO'
  | 'RR'
  | 'RS'
  | 'SC'
  | 'SE'
  | 'SP'
  | 'TO'
  | 'EX';
export type TAmb = '1' | '2';
export type Tmod = '55' | '65';
type TNFeModBC = '0' | '1' | '2' | '3';
type TNFeModBCST = '0' | '1' | '2' | '3' | '4' | '5';
type TConsStatServXServ = 'STATUS';
type TEnviNFeIndSinc = '0' | '1';
type TNFeInfNFeIdeIndIntermed = '0' | '1' | '2' | '3' | '4' | '9';
// ...
type TNFeInfNFeEmitCRT = '1' | '2' | '3' | '4';
type ItemsChoiceType5 = 'balsa' | 'reboque' | 'vagao' | 'veicTransp';
// ...
type TNFeDoc = 'CNPJ' | 'CPF';
// ...
export interface TNFeInfNFeDet {
  $: {
    nItem: string;
  };
  prod: TNFeInfNFeDetProd;
  imposto: TNFeInfNFeDetImposto;
  impostoDevol?: TNFeInfNFeDetImpostoDevol;
  infAdProd?: string;
}
interface TNFeInfNFeDetImpostoDevol {
  pDevol: number;
  IPI: TNFeInfNFeDetImpostoDevolIPI;
}
// ...
interface TNFeInfNFeDetImposto {
  vTotTrib?: number;
  ICMS: TNFeInfNFeDetImpostoICMS;
  PIS: TNFeInfNFeDetImpostoPIS;
  COFINS: TNFeInfNFeDetImpostoCOFINS;
  IPI?: TNFeInfNFeDetImpostoIPI;
  II?: TNFeInfNFeDetImpostoII;
  ISSQN?: TNFeInfNFeDetImpostoISSQN;
  PISST?: TNFeInfNFeDetImpostoPISST;
  COFINSST?: TNFeInfNFeDetImpostoCOFINSST;
  ICMSUFDest?: TNFeInfNFeDetImpostoICMSUFDest;
}
interface TNFeInfNFeDetImpostoICMSUFDest {
  vBCUFDest: number;
  vBCFCPUFDest: number;
  pFCPUFDest: number;
  pICMSUFDest: number;
  pICMSInter: number;
  pICMSInterPart: number;
  vFCPUFDest: number;
  vICMSUFDest: number;
  vICMSUFRemet: number;
}
// ...
interface TNFeInfNFeDetProd {
  cProd: string;
  cEAN: string;
  xProd: string;
  NCM: string;
  nVE?: string[];
  CEST?: string;
  indEscala?: TNFeInfNFeDetProdIndEscala;
  indEscalaSpecified?: boolean;
  cNPJFab?: string;
  cBenef?: string;
  eXTIPI?: string;
  CFOP: string;
  uCom: string;
  qCom: string;
  vUnCom: string;
  vProd: string;
  cEANTrib: string;
  uTrib: string;
  qTrib: string;
  vUnTrib: string;
  vFrete: string;
  vSeg: string;
  vDesc: string;
  vOutro: string;
  indTot: '0' | '1';
  di?: TNFeInfNFeDetProdDI[];
  det?: TNFeInfNFeDetProdDet[];
  xPed?: string;
  nItemPed?: string;
  nFCI?: string;
  rastro?: TNFeInfNFeDetProdRastro[];
  comb?: TNFeInfNFeDetProdComb;
  items?: object[];
}
// ...
export interface TNFeInfNFeDest {
  CNPJ?: string;
  CPF?: string;
  idEstrangeiro?: string;
  xNome: string;
  enderDest?: TEnderEmi;
  indIEDest: '1' | '2' | '9';
  IE?: string;
  ISUF?: string;
  IM?: string;
  email?: string;
}
interface TEnderEmi {
  xLgr: string;
  nro: string;
  xCpl?: string;
  xBairro: string;
  cMun: string;
  xMun: string;
  UF: TUfEmi;
  CEP: string;
  cPais: TEnderEmiCPais;
  cPaisSpecified?: boolean;
  xPais: TEnderEmiXPais;
  xPaisSpecified?: boolean;
  fone?: string;
}
export interface TNFeInfNFeEmit {
  CNPJ: string;
  xNome: string;
  xFant: string;
  enderEmit: TEnderEmi;
  fone?: string;
  IE: string;
  iEST?: string;
  IM: string;
  CNAE?: string;
  CRT: TNFeInfNFeEmitCRT;
}
export interface TNFeInfNFeTransp {
  modFrete: '0' | '1' | '2' | '3' | '4' | '9';
  transporta?: TNFeInfNFeTranspTransporta;
  retTransp?: TNFeInfNFeTranspRetTransp;
  veicTransp?: TNFeInfNFeTranspVeicTransp;
  items?: object[];
  itemsElementName?: ItemsChoiceType5[];
  vol?: TNFeInfNFeTranspVol[];
}
interface TNFeInfNFeTranspVol {
  qVol: string;
  esp: string;
  marca?: string;
  nVol?: string;
  pesoL: string;
  pesoB: string;
  lacres?: TNFeInfNFeTranspVolLacres[];
}
interface TNFeInfNFeTranspVolLacres {
  nLacre: string;
}
interface TNFeInfNFeTranspVeicTransp {
  placa: string;
  UF: string;
  RNTC: string;
}
interface TNFeInfNFeTranspRetTransp {
  vServ: string;
  vBCRet: string;
  pICMSRet: string;
  vICMSRet: string;
  cFOP: string;
  cMunFG: string;
}
interface TNFeInfNFeTranspTransporta {
  CNPJ?: string;
  CPF?: string;
  item: string;
  itemElementName: TNFeDoc;
  xNome: string;
  ie: string;
  xEnder: string;
  xMun: string;
  uf: TUfEmi;
  ufSpecified: boolean;
}
export interface TNFeInfNFePag {
  detPag: TNFeInfNFePagDetPag[];
  vTroco?: string;
}
export interface TNFeInfNFePagDetPag {
  indPag: '0' | '1';
  indPagSpecified?: boolean;
  tPag: '01' | '02' | '03' | '04' | '05' | '10' | '11' | '12' | '13' | '14' | '15' | '16' | '17' | '18' | '19' | '90' | '99';
  vPag: string;
  card?: TNFeInfNFePagDetPagCard;
  xPag?: string;
}
interface TNFeInfNFePagDetPagCard {
  tpIntegra: TNFeInfNFePagDetPagCardTpIntegra;
  CNPJ: string;
  tBand: TNFeInfNFePagDetPagCardTBand;
  tBandSpecified?: boolean;
  cAut: string;
}
export interface TNFeInfNFeInfAdic {
  infAdFisco?: string;
  infCpl?: string;
  obsCont: TNFeInfNFeInfAdicObs[];
  obsFisco: TNFeInfNFeInfAdicObs[];
  procRef: TNFeInfNFeInfAdicProcRef[];
}
export interface TNFeInfNFeInfAdicObs {
  xTexto: string;
  xCampo: string;
}
export interface TNFeInfNFeInfAdicProcRef {
  nProc: string;
  indProc: TNFeInfNFeInfAdicProcRefIndProc;
}
export interface TNFeInfNFeTotalICMSTot {
  vBC: string;
  vICMS: string;
  vICMSDeson: string;
  vFCPUFDest?: string;
  vICMSUFDest?: string;
  vICMSUFRemet?: string;
  vFCP: string;
  vBCST: string;
  vST: string;
  vFCPST: string;
  vFCPSTRet: string;
  qBCMono?: string;
  vICMSMono?: string;
  qBCMonoReten?: string;
  vICMSMonoReten?: string;
  qBCMonoRetet?: string;
  vICMSMonoRet?: string;
  vProd: string;
  vFrete: string;
  vSeg: string;
  vDesc: string;
  vII: string;
  vIPI: string;
  vIPIDevol: string;
  vPIS: string;
  vCOFINS: string;
  vOutro: string;
  vNF: string;
  vTotTrib?: string;
}
// ...
interface TNFeInfNFeIdeNFref {
  refCTe?: string;
  refECF?: TNFeInfNFeIdeNFrefRefECF;
  refNF?: TNFeInfNFeIdeNFrefRefNF;
  refNFP?: TNFeInfNFeIdeNFrefRefNFP;
  refNFe?: string;
}
// ...
export interface TNFeInfNFeIde {
  cUF: TCodUfIbge;
  cNF: string;
  natOp: string;
  mod: Tmod;
  serie: string;
  nNF: string;
  dhEmi: string;
  dhSaiEnt?: string;
  tpNF: '0' | '1';
  idDest: '1' | '2' | '3';
  cMunFG: string;
  tpImp: '1' | '2' | '3' | '4' | '5';
  tpEmis: '1' | '2' | '3' | '4' | '5' | '6' | '7' | '9';
  cDV: string;
  tpAmb: TAmb;
  finNFe: '1' | '2' | '3' | '4';
  indFinal: '0' | '1';
  indPres: '1' | '2' | '3' | '4' | '5' | '9';
  procEmi: '0' | '1' | '2' | '3';
  verProc: string;
  dhCont?: string;
  xJust?: string;
  NFref?: TNFeInfNFeIdeNFref[];
  indIntermed?: TNFeInfNFeIdeIndIntermed;
}
export interface TNFeInfNFeTotal {
  ICMSTot: TNFeInfNFeTotalICMSTot;
  ISSQNtot?: TNFeInfNFeTotalISSQNtot;
  retTrib?: TNFeInfNFeTotalRetTrib;
}
interface TNFeInfNFeTotalRetTrib {
  vRetPIS: string;
  vRetCOFINS: string;
  vRetCSLL: string;
  vBCIRRF: string;
  vIRRF: string;
  vBCRetPrev: string;
  vRetPrev: string;
}
interface TNFeInfNFeTotalISSQNtot {
  vServ: string;
  vBC: string;
  vISS: string;
  vPIS: string;
  vCOFINS: string;
  dCompet: string;
  vDeducao: string;
  vOutro: string;
  vDescIncond: string;
  vDescCond: string;
  vISSRet: string;
  cRegTrib: TNFeInfNFeTotalISSQNtotCRegTrib;
  cRegTribSpecified: boolean;
}
export interface TNFeInfNFeCobr {
  fat: TNFeInfNFeCobrFat;
  dup: TNFeInfNFeCobrDup[];
}
export interface TNFeInfNFeCobrDup {
  nDup: string;
  dVenc: string;
  vDup: string;
}
interface TNFeInfNFeCobrFat {
  nFat: string;
  vOrig: string;
  vDesc: string;
  vLiq: string;
}
export interface TNFeInfNFe {
  nfe?: TNFeInfNFeDetImpostoICMSUFDest;
  $: {
    versao: string;
    Id: string;
  };
  ide: TNFeInfNFeIde;
  emit: TNFeInfNFeEmit;
  avulsa?: TNFeInfNFeAvulsa;
  dest?: TNFeInfNFeDest;
  retirada?: TLocal;
  entrega?: TLocal;
  autXML?: TNFeInfNFeAutXML[];
  det: TNFeInfNFeDet[];
  total: TNFeInfNFeTotal;
  transp: TNFeInfNFeTransp;
  cobr?: TNFeInfNFeCobr;
  pag?: TNFeInfNFePag;
  infAdic: TNFeInfNFeInfAdic;
  a?: TNFeInfNFea;
  compra?: TNFeInfNFeCompra;
  cana?: TNFeInfNFeCana;
  infRespTec?: TInfRespTec;
}
export interface TInfRespTec {
  CNPJ: string;
  xContato: string;
  email: string;
  fone: string;
  idCSRT?: string;
  hashCSRT?: string;
}
interface TNFeInfNFeCana {
  safra: string;
  ref: string;
  forDia: TNFeInfNFeCanaForDia[];
  qTotMes: string;
  qTotAnt: string;
  qTotGer: string;
  deduc: TNFeInfNFeCanaDeduc[];
  vFor: string;
  vTotDed: string;
  vLiqFor: string;
}
interface TNFeInfNFeCanaDeduc {
  xDed: string;
  vDed: string;
}
interface TNFeInfNFeCanaForDia {
  qtde: string;
  dia: string;
}
interface TNFeInfNFeCompra {
  xNEmp: string;
  xPed: string;
  xCont: string;
}
interface TNFeInfNFea {
  uFSaidaPais: TUfEmi;
  xLoca: string;
  xLocDespacho: string;
}
interface TNFeInfNFeAutXML {
  item: string;
  itemElementName: TNFeDoc;
}
interface TLocal {
  item: string;
  itemElementName: TNFeDoc;
  xNome: string;
  xLgr: string;
  nro: string;
  xCpl: string;
  xBairro: string;
  cMun: string;
  xMun: string;
  uf: TUfEmi;
  cEP: string;
  cPais: string;
  xPais: string;
  fone: string;
  email: string;
  ie: string;
}
interface TNFeInfNFeAvulsa {
  cNPJ: string;
  xOrgao: string;
  matr: string;
  xAgente: string;
  fone: string;
  uf: TUfEmi;
  nDAR: string;
  dEmi: string;
  vDAR: string;
  repEmi: string;
  dPag: string;
}